refactor(admissions): RegisterInformationForm API and Logic layers
Convert top layers to use hooks instead and rename the view
layer Form for improved API
